LRAC-10636 rename module to analytics-dxp-entity-api and analytics-dxp-entity-impl
