--- modules/apps/analytics/analytics-dxp-entities-exporter-api/src/main/java/com/liferay/analytics/dxp/entities/exporter/retriever/DXPEntityRetriever.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.analytics.dxp.entities.exporter.retriever;
-
-import com.liferay.analytics.dxp.entities.exporter.dto.v1_0.DXPEntity;
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.portal.kernel.model.BaseModel;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-
-/**
- * @author Marcos Martins
- */
-public interface DXPEntityRetriever {
-
-	public Page<DXPEntity> getDXPEntitiesPage(
-			long companyId, Pagination pagination,
-			UnsafeFunction<BaseModel<?>, DXPEntity, Exception>
-				transformUnsafeFunction)
-		throws Exception;
-
-}
