--- modules/apps/analytics/analytics-dxp-entities-exporter-impl/rest-openapi.yaml
@@ -1,42 +1,0 @@
-components:
-    schemas:
-        DXPEntity:
-            properties:
-                expandoFields:
-                    items:
-                        $ref: "#/components/schemas/ExpandoField"
-                    type: array
-                fields:
-                    items:
-                        $ref: "#/components/schemas/Field"
-                    type: array
-                id:
-                    type: string
-                type:
-                    type: string
-            type: object
-        ExpandoField:
-            properties:
-                columnId:
-                    format: int64
-                    type: integer
-                name:
-                    type: string
-                value:
-                    type: string
-            type: object
-        Field:
-            properties:
-                name:
-                    type: string
-                value:
-                    type: object
-            type: object
-info:
-    license:
-        name: "Apache 2.0"
-        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
-    title: Liferay Analytics DXP Entities Exporter
-    version: v1.0
-openapi: 3.0.1
-paths:
