+++ modules/apps/analytics/analytics-dxp-entity-api/src/main/resources/com/liferay/analytics/dxp/entity/dto/v1_0/packageinfo
