+++ modules/apps/analytics/analytics-dxp-entity-impl/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Analytics DXP Entity Implementation
+Bundle-SymbolicName: com.liferay.analytics.dxp.entity.impl
+Bundle-Version: 1.0.0
+-dsannotations-options: inherit
