+++ modules/apps/analytics/analytics-dxp-entity-impl/build.gradle
@@ -1,0 +1,18 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "io.swagger.core.v3", name: "swagger-annotations", version: "2.0.5"
+	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.1"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
+	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:analytics:analytics-batch-export-import-api")
+	compileOnly project(":apps:analytics:analytics-dxp-entity-api")
+	compileOnly project(":apps:analytics:analytics-message-sender-api")
+	compileOnly project(":apps:analytics:analytics-settings-api")
+	compileOnly project(":apps:batch-engine:batch-engine-api")
+	compileOnly project(":apps:dispatch:dispatch-api")
+	compileOnly project(":apps:portal-odata:portal-odata-api")
+	compileOnly project(":apps:portal-vulcan:portal-vulcan-api")
+	compileOnly project(":core:osgi-service-tracker-collections")
+	compileOnly project(":core:petra:petra-function")
+	compileOnly project(":core:petra:petra-string")
+}
