+++ modules/apps/analytics/analytics-dxp-entity-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/openapi.properties
